Complete to diploma
Add project type CRUD
Add small fixes to existing code
Add CreateUpdateProjectRequest for practicing request creating
Change database schema default string length from 191 to 256
Change project index view from list-like to table-like
Add sorting by one column on the project index view
Add 'registration_number', 'registered_at', 'defended_at' and 'grade'
to project model

+++ app/Http/Controllers/SupervisorController.php
@@ -11,14 +11,14 @@ class SupervisorController extends Controller
 {
     public function index()
     {
-        $supervisors = Supervisor::all();
+        $supervisors = Supervisor::query()->withCount('projects')->paginate(10);
         
         return view('supervisors.index', compact('supervisors'));
     }
     
     public function create()
     {
-        return view('supervisors.index');
+        return view('supervisors.create');
     }
     
     public function store(Request $request): RedirectResponse
@@ -34,7 +34,7 @@ class SupervisorController extends Controller
         
         Supervisor::query()->create(['name' => $request->input('name')]);
         
-        return back()->with('message', 'Керівник успішно створений.');
+        return back()->with('message', 'Керівник успішно доданий.');
     }
     
     public function edit(Supervisor $supervisor)
@@ -48,7 +48,7 @@ class SupervisorController extends Controller
             'name' => [
                 'required',
                 'string',
-                'between:3,10',
+                'between:3,255',
                 Rule::unique('groups', 'name')
                     ->ignore($supervisor->name, 'name'),
             ],
@@ -61,6 +61,12 @@ class SupervisorController extends Controller
     
     public function destroy(Supervisor $supervisor): RedirectResponse
     {
+        if ($supervisor->projects()->count() < 0) {
+            return back()->with([
+                'message' => 'Спочатку видаліть чи відредагуйте усі проекти з даним керівником',
+            ]);
+        }
+        
         $supervisor->delete();
         
         return back();
